Move Home to component folder
