+++ assets/js/app.js
@@ -3,6 +3,7 @@ import React from "react"
 import ReactDOM from "react-dom"
 import { BrowserRouter as Router, Route, Link } from 'react-router-dom'
 import axios from "axios"
+import Home from './components/home'
 
 class HelloReact extends React.Component {
   render() {
@@ -15,22 +16,6 @@ class HelloReact extends React.Component {
         <Route exact path="/books/new" component={NewBook}/>
       </div>
       </Router>
-    )
-  }
-}
-
-class Home extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Welcome to PickABook!</h1>
-        <Router>
-          <Link to="/books">All books</Link>
-        </Router>
-        <Router>
-          <Link to="/books/new">Add a book</Link>
-        </Router>
-      </div>
     )
   }
 }
@@ -174,4 +159,3 @@ class Book extends React.Component {
 // )
 
 module.exports = HelloReact;
-module.exports = Home;
